+++ smarts/env/gymnasium/driving_smarts_2023_env.py
@@ -162,7 +162,7 @@ def resolve_agent_interface(agent_interface: AgentInterface):
     )
     max_episode_steps = 1000
     waypoints_lookahead = 80
-    neighborhood_radius = 50
+    neighborhood_radius = 80
     return AgentInterface(
         accelerometer=True,
         action=agent_interface.action,