+++ smarts/env/gymnasium/platoon_env.py
@@ -169,7 +169,7 @@ def resolve_agent_interface(agent_interface: AgentInterface):
     )
     max_episode_steps = 1000
     waypoints_lookahead = 80
-    neighborhood_radius = 50
+    neighborhood_radius = 80
     return AgentInterface(
         accelerometer=True,
         action=agent_interface.action,