+++ CHANGELOG.md
@@ -26,7 +26,8 @@ Copy and pasting the git commit messages is __NOT__ enough.
 - Added engine `observation_workers` configuration which can be used to configure the number of parallel sensor workers: 0 runs the sensors on the local thread, >=1 runs using the multiprocessing backing.
 - Added engine `sensor_parallelization` configuration of sensor parallelization backing, options ("mp"|"ray"): "mp" python multiprocessing, "ray" ray worker backing.
 - Added engine `reset_retries` configuration engine retries before the simulator will raise an error on reset.
-- Introduced new comfort metric in metric cost functions.
+- Introduced new comfort cost function in metric module.
+- Introduced new gap-between-vehicles cost function in metric module.
 ### Changed
 - The trap manager, `TrapManager`, is now a subclass of `ActorCaptureManager`.
 - Considering lane-change time ranges between 3s and 6s, assuming a speed of 13.89m/s, the via sensor lane acquisition range was increased from 40m to 80m, for better driving ability.
@@ -48,6 +49,7 @@ Copy and pasting the git commit messages is __NOT__ enough.
 - In the metrics module, the records which is the raw metrics data and the scoring which is the formula to compute the final results are now separated to provided greater flexibility for applying metrics to different environments.
 - Benchmark listing may specify specialised metric formula for each benchmark.
 - Changed `benchmark_runner_v0.py` to only average records across scenarios that share the same environment. Records are not averaged across different environments, because the scoring formula may differ in different environments.
+- Renamed GapBetweenVehicles cost to VehicleGap cost in metric module.
 ### Deprecated
 ### Fixed
 - Fixed issues related to waypoints in junctions on Argoverse maps. Waypoints will now be generated for all paths leading through the lane(s) the vehicle is on.